style: clear workflow service clippy warnings
Clean up existing workflow-service clippy warnings found while running the Stage 11 Wave 02 package gate.

Use clearer control flow in trace projection, remove an unnecessary cast, switch bool comparisons to assert!, and scope clippy argument-count allowances to existing scheduler diagnostic helper functions.

Verification:

- cargo clippy -p pantograph-workflow-service --all-targets -- -D warnings

+++ crates/pantograph-workflow-service/src/workflow/session_execution_api.rs
@@ -254,8 +254,8 @@ impl WorkflowService {
                         );
                     }
                     if !runtime_admission_delay_recorded {
-                        let delayed_until_ms = unix_timestamp_ms()
-                            .saturating_add(WORKFLOW_SESSION_QUEUE_POLL_MS as u64);
+                        let delayed_until_ms =
+                            unix_timestamp_ms().saturating_add(WORKFLOW_SESSION_QUEUE_POLL_MS);
                         self.record_scheduler_delay_event_if_configured(
                             &session,
                             run_snapshot.as_ref(),
@@ -900,6 +900,7 @@ impl WorkflowService {
         .map_err(WorkflowServiceError::from)
     }
 
+    #[allow(clippy::too_many_arguments)]
     fn record_scheduler_delay_event_if_configured(
         &self,
         session: &WorkflowExecutionSessionSummary,
@@ -1162,6 +1163,7 @@ impl WorkflowService {
         .map_err(WorkflowServiceError::from)
     }
 
+    #[allow(clippy::too_many_arguments)]
     fn record_scheduler_reservation_event_if_configured(
         &self,
         session: &WorkflowExecutionSessionSummary,
